+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# Changelog for Unicode Guards v0.2.0
+
+This is the changelog for Unicode Guards v0.2.0 released on March 11th, 2020.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-unicode/unicode_guards/tags)
+
+## Enhancements
+
+* Add guards for quote marks. Adds:
+
+  * `is_quote_mark/1`
+  * `is_quote_mark_left/1`
+  * `is_quote_mark_right/1`
+  * `is_quote_mark_ambidextrous/1`
+  * `is_quote_mark_single/1`
+  * `is_quote_mark_double/1`
+
 # Changelog for Unicode Guards v0.1.1
 
 This is the changelog for Unicode Guards v0.1.1 released on February 25th, 2020.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-unicode/unicode_guards/tags)